base: remove support for power off Alarm
This feature is currently causing a constaint I/O load.
It doesn't seem like there would be a possibility to avoid it.
-> Globally remove the frameworks support because of noticed issues
   and the overall non usefulness of the feature.

It is suggested that the feature will be disabled in the
kernels too.

Reverted commits:

Revert "Add support for power off alarm"

This reverts commit 6a4dfa59ebd17206b596e56d9a794fce75c36b2d.

Revert "Alarm: Add one more RTC alarm type for poweroff alarm."

This reverts commit ed29480f3ce6176ee363cedcb8f4ae2fa51f3e5d.

Change-Id: I6f83d6115a2f6cfc25afc01c670e37270f5dcd23
Signed-off-by: Alex Naidis <alex.naidis@linux.com>